forcing lb and ub to be float, otherwise njit throws

+++ src/vpso/initialization.py
@@ -77,8 +77,8 @@ def adjust_dimensions(
         Returns the `lb`, `ub`, `nvec`, `dim`, `max_velocity_rate`, `w`, `c1`, `c2`,
         `ftol`, `xtol`, and `patience`.
     """
-    lb = np.expand_dims(lb, 1)  # add swarm dimension
-    ub = np.expand_dims(ub, 1)  # add swarm dimension
+    lb = np.expand_dims(lb, 1).astype(float, copy=True)  # add swarm dimension
+    ub = np.expand_dims(ub, 1).astype(float, copy=True)  # add swarm dimension
     nvec, _, dim = lb.shape
     return (
         lb,
